Switched to long instead of int

+++ VamosVamosServer/DAO/Implementation/MatchDAO.cs
@@ -10,7 +10,7 @@ namespace VamosVamosServer.DAO.Implementation;
 
 public class MatchDAO : IMatchDAO
 {
-    public ActionResult UpdateMatchScore(string token, int matchId, int scoreTeam1, int scoreTeam2)
+    public ActionResult UpdateMatchScore(string token, long matchId, long scoreTeam1, long scoreTeam2)
     {
         UserDAO userDao = new UserDAO();
         ActionResult r = userDao.GetUserByToken(token);
@@ -61,7 +61,7 @@ public class MatchDAO : IMatchDAO
         return new OkResult();
     }
 
-    public ActionResult<Match?> GetMatch(string token, int matchId)
+    public ActionResult<Match?> GetMatch(string token, long matchId)
     {
         UserDAO userDao = new UserDAO();
         ActionResult r = userDao.GetUserByToken(token);
@@ -141,7 +141,7 @@ public class MatchDAO : IMatchDAO
         return new OkObjectResult(match);
     }
 
-    public ActionResult<List<Match>> GetAllMatches(string token, int page, int limit)
+    public ActionResult<List<Match>> GetAllMatches(string token, long page, long limit)
     {
         UserDAO userDao = new UserDAO();
         ActionResult r = userDao.GetUserByToken(token);
