+++ VamosVamosServer/Model/PlayedModel/IAssist.cs
@@ -2,7 +2,7 @@ namespace VamosVamosServer.Model.PlayedModel;
 
 public interface IAssist
 {
-    int AssistTime { get; set; }
+    long AssistTime { get; set; }
     AssistType AssistType { get; set; }
     Played Played { get; set; }
     Played AssistedPlayed { get; set; }
