+++ VamosVamosServer/Model/PlayedModel/IPenalty.cs
@@ -2,7 +2,7 @@ namespace VamosVamosServer.Model.PlayedModel;
 
 public interface IPenalty
 {
-    int PenaltyTime { get; set; }
+    long PenaltyTime { get; set; }
     PenaltyResult Result { get; set; }
     PenaltyObtainingMethod ObtainingMethod { get; set; }
     Played Played { get; set; }
