+++ VamosVamosServer/Service/Interfaces/IMatchService.cs
@@ -12,7 +12,7 @@ public interface IMatchService
     /// <param name="matchId">The match id</param>
     /// <param name="scoreTeam1">The new team1 score</param>
     /// <param name="scoreTeam2">The new team2 score</param>
-    ActionResult UpdateMatchScore(string token, int matchId, int scoreTeam1, int scoreTeam2);
+    ActionResult UpdateMatchScore(string token, long matchId, long scoreTeam1, long scoreTeam2);
 
     /// <summary>
     /// Get a match by its id.
@@ -20,7 +20,7 @@ public interface IMatchService
     /// <param name="token">The user's token</param>
     /// <param name="matchId">The match id</param>
     /// <returns>The match</returns>
-    ActionResult<Match?> GetMatch(string token, int matchId);
+    ActionResult<Match?> GetMatch(string token, long matchId);
 
     /// <summary>
     /// Get all matches by page.
@@ -29,5 +29,5 @@ public interface IMatchService
     /// <param name="page">The page number</param>
     /// <param name="limit">The number of matches per page</param>
     /// <returns>All the matches of the page</returns>
-    ActionResult<List<Match>> GetAllMatches(string token, int page, int limit);
+    ActionResult<List<Match>> GetAllMatches(string token, long page, long limit);
 }
